requirements.txt
langgraph>=0.2.60
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
langsmith>=0.1.100
python-telegram-bot>=21.0
yfinance>=0.2.50
ddgs>=0.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
